decimal places not considered while setting outer box number values on flow optimization. [Issue: 39701]
git-svn-id: http://dev.farrms.us:18080/pioneer/products/trmtracker/branches/TRMTracker_Enercity/Ver_4.2.003@117573 fd52c265-6bb1-e54c-ad91-b501145a01df

+++ FARRMS/trm/adiha.html.forms/_scheduling_delivery/gas/flow_optimization/hourly.scheduling.php
@@ -430,8 +430,8 @@ echo $sch_obj->close_layout();
         var first_hour_del_vol = 0;
         subgrid.forEachCell(subgrid.getRowId(rec_row_index), function(cellObj, cid) {
             if (cid > 6) { //only for hour columns
-                var rec_hrly = parseInt(cellObj.getValue() == '' ? 0 : cellObj.getValue()); 
-                var del_hrly = parseInt(subgrid.cells2(del_row_index, cid).getValue() == '' ? 0 : subgrid.cells2(del_row_index, cid).getValue());
+                var rec_hrly = parseFloat(cellObj.getValue() == '' ? 0 : cellObj.getValue()); 
+                var del_hrly = parseFloat(subgrid.cells2(del_row_index, cid).getValue() == '' ? 0 : subgrid.cells2(del_row_index, cid).getValue());
 
                 if (cid == 7) {
                     first_hour_rec_vol = rec_hrly;
@@ -440,7 +440,7 @@ echo $sch_obj->close_layout();
 
                 total_rec += rec_hrly;
                 total_del += del_hrly;
-                total_path_rmdq += parseInt(getNumberFormat(subgrid.cells2(0, cid).getValue().split('/')[1], '', 1));
+                total_path_rmdq += parseFloat(getNumberFormat(subgrid.cells2(0, cid).getValue().split('/')[1], '', 1));
             }
         });
